Perbaikan

+++ api/models/Post.go
@@ -10,7 +10,7 @@ import (
 )
 
 type Post struct {
-	ID        uint      `gorm:"primary_key;auto_increment" json:"id"`
+	ID        uint64    `gorm:"primary_key;auto_increment" json:"id"`
 	Phone     string    `gorm:"size:255;not null;" json:"phone"`
 	Content   string    `gorm:"text;not null;" json:"content"`
 	Author    Admin     `json:"author"`
